pkp/pkp-lib#10616 Add ThemePlugin function to get localized theme option value
